// app/src/bp-api/models/DustConversion.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DustConversion = {
  /**
   * The ID of the dust conversion.
   */
  id: number
  /**
   * Dust quantity.
   */
  quantity: string
  /**
   * Symbol.
   */
  symbol: string
  /**
   * USDC received.
   */
  usdcReceived: string
  /**
   * Timestamp.
   */
  timestamp: string
}