// app/src/bp-api/models/PnlPayment.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PnlPayment = {
  /**
   * Quantity of the payment (`USDC`).
   */
  pnlRealized: string
  /**
   * The market symbol the payment is associated with.
   */
  symbol: string
  /**
   * The timestamp of the payment (UTC).
   */
  timestamp: string
}